+++ src/+groundwater/calculateGroundWaterDepth.m
@@ -6,8 +6,5 @@ function groundWaterDepth = calculateGroundWaterDepth(topLevel, headBotmLayer, T
     if groundWaterDepth <= 0
         warning('The soil is fully saturated up to the land surface level!');
         groundWaterDepth = 1.0; % to avoid model crashing, assign minimum groundWaterDepth value of 1 cm
-    elseif groundWaterDepth > Tot_Depth
-        warning('Groundwater table is below the end of the soil column!');
     end
-
 end