+++ CHANGELOG.md
@@ -7,6 +7,13 @@
   [#234](https://github.com/EcoExtreML/STEMMUS_SCOPE/pull/234)
 - Save water stress factor and water potential into csv files.
   [#229](https://github.com/EcoExtreML/STEMMUS_SCOPE/pull/229)
+- Add an option to define soil layer settings through a csv file discussed in
+  [#237](https://github.com/EcoExtreML/STEMMUS_SCOPE/issues/237) and
+  [#241](https://github.com/EcoExtreML/STEMMUS_SCOPE/issues/241) and added in
+  [#243](https://github.com/EcoExtreML/STEMMUS_SCOPE/pull/243)
+- Remove unnecessary function discussed in
+  [#244](https://github.com/EcoExtreML/STEMMUS_SCOPE/issues/244) and removed in
+  [#243](https://github.com/EcoExtreML/STEMMUS_SCOPE/pull/243)
 
 **Fixed:**
 
@@ -22,8 +29,7 @@
 - Defining the indices of the first four layers discussed in
   [#237](https://github.com/EcoExtreML/STEMMUS_SCOPE/issues/237) and fixed in
   [#238](https://github.com/EcoExtreML/STEMMUS_SCOPE/pull/238)
-
-
+  
 <a name="1.5.0"></a>
 # [1.5.0](https://github.com/EcoExtreML/STEMMUS_SCOPE/releases/tag/1.5.0) - 3 Jan 2024